Separate out two layouts again
