+++ dotcom-rendering/src/components/SubMeta.tsx
@@ -16,7 +16,6 @@ import type { BaseLinkType } from '../model/extract-nav';
 import { palette } from '../palette';
 import { Island } from './Island';
 import { ShareButton } from './ShareButton.island';
-import { interactiveLayoutSwitchoverDate } from '../layouts/DecideLayout';
 
 const labelStyles = (design: ArticleDesign): SerializedStyles => css`
 	${design === ArticleDesign.Gallery ? grid.column.centre : undefined};
@@ -131,6 +130,7 @@ type Props = {
 	webUrl: string;
 	webTitle: string;
 	showBottomSocialButtons: boolean;
+	isDeprecatedInteractiveLayout?: boolean;
 };
 
 const syndicationButtonOverrides = css`
@@ -206,6 +206,7 @@ export const SubMeta = ({
 	webUrl,
 	webTitle,
 	showBottomSocialButtons,
+	isDeprecatedInteractiveLayout = false,
 }: Props) => {
 	const createLinks = () => {
 		const links: BaseLinkType[] = [];
@@ -226,15 +227,11 @@ export const SubMeta = ({
 		format.design !== ArticleDesign.Interactive &&
 		format.design !== ArticleDesign.Gallery;
 
-	const usesDeprecatedInteractiveLayout =
-		format.design === ArticleDesign.Interactive &&
-		interactiveLayoutSwitchoverDate > new Date();
-
 	return (
 		<div
 			data-print-layout="hide"
 			css={[
-				usesDeprecatedInteractiveLayout ? setMetaWidth : undefined,
+				isDeprecatedInteractiveLayout ? setMetaWidth : undefined,
 				format.design === ArticleDesign.Gallery
 					? galleryStyles
 					: bottomPadding,
