+++ dotcom-rendering/src/layouts/DecideLayout.tsx
@@ -15,6 +15,7 @@ import { HostedGalleryLayout } from './HostedGalleryLayout';
 import { HostedVideoLayout } from './HostedVideoLayout';
 import { ImmersiveLayout } from './ImmersiveLayout';
 import { InteractiveLayout } from './InteractiveLayout';
+import { InteractiveLayoutDeprecated } from './InteractiveLayoutDeprecated';
 import { LiveLayout } from './LiveLayout';
 import { NewsletterSignupLayout } from './NewsletterSignupLayout';
 import { PictureLayout } from './PictureLayout';
@@ -140,17 +141,24 @@ const DecideLayoutApps = ({ article, renderingTarget }: AppProps) => {
 		default: {
 			switch (article.design) {
 				case ArticleDesign.Interactive:
-					return (
-						<InteractiveLayout
-							article={article.frontendData}
-							format={format}
-							renderingTarget={renderingTarget}
-							useDeprecatedGrid={
-								publicationDate <
-								interactiveLayoutSwitchoverDate
-							}
-						/>
-					);
+					if (publicationDate < interactiveLayoutSwitchoverDate) {
+						return (
+							<InteractiveLayoutDeprecated
+								article={article.frontendData}
+								format={format}
+								renderingTarget={renderingTarget}
+								serverTime={serverTime}
+							/>
+						);
+					} else {
+						return (
+							<InteractiveLayout
+								article={article.frontendData}
+								format={format}
+								renderingTarget={renderingTarget}
+							/>
+						);
+					}
 				case ArticleDesign.FullPageInteractive: {
 					return (
 						<FullPageInteractiveLayout
@@ -334,18 +342,26 @@ const DecideLayoutWeb = ({ article, NAV, renderingTarget }: WebProps) => {
 		default: {
 			switch (article.design) {
 				case ArticleDesign.Interactive:
-					return (
-						<InteractiveLayout
-							article={article.frontendData}
-							NAV={NAV}
-							format={format}
-							renderingTarget={renderingTarget}
-							useDeprecatedGrid={
-								publicationDate <
-								interactiveLayoutSwitchoverDate
-							}
-						/>
-					);
+					if (publicationDate < interactiveLayoutSwitchoverDate) {
+						return (
+							<InteractiveLayoutDeprecated
+								article={article.frontendData}
+								NAV={NAV}
+								format={format}
+								renderingTarget={renderingTarget}
+								serverTime={serverTime}
+							/>
+						);
+					} else {
+						return (
+							<InteractiveLayout
+								article={article.frontendData}
+								NAV={NAV}
+								format={format}
+								renderingTarget={renderingTarget}
+							/>
+						);
+					}
 				case ArticleDesign.FullPageInteractive: {
 					return (
 						<FullPageInteractiveLayout
