+++ dotcom-rendering/src/layouts/InteractiveLayoutDeprecated.tsx
@@ -2,14 +2,12 @@ import { css, Global } from '@emotion/react';
 import {
 	from,
 	palette as sourcePalette,
-	space,
 	until,
 } from '@guardian/source/foundations';
 import { Hide } from '@guardian/source/react-components';
 import { StraightLines } from '@guardian/source-development-kitchen/react-components';
 import type React from 'react';
 import { AdSlot, MobileStickyContainer } from '../components/AdSlot.web';
-import { AffiliateDisclaimer } from '../components/AffiliateDisclaimer';
 import { AppsFooter } from '../components/AppsFooter.island';
 import { ArticleBody } from '../components/ArticleBody';
 import { ArticleContainer } from '../components/ArticleContainer';
@@ -20,7 +18,7 @@ import { ArticleTitle } from '../components/ArticleTitle';
 import { Border } from '../components/Border';
 import { Carousel } from '../components/Carousel.island';
 import { DecideLines } from '../components/DecideLines';
-import { DirectoryPageNavIsland } from '../components/DirectoryPageNavIsland';
+import { DirectoryPageNav } from '../components/DirectoryPageNav';
 import { DiscussionLayout } from '../components/DiscussionLayout';
 import { Footer } from '../components/Footer';
 import { GridItem } from '../components/GridItem';
@@ -206,7 +204,7 @@ interface AppsProps extends CommonProps {
 	renderingTarget: 'Apps';
 }
 
-export const InteractiveLayout = (props: WebProps | AppsProps) => {
+export const InteractiveLayoutDeprecated = (props: WebProps | AppsProps) => {
 	const { article, format, renderingTarget, serverTime } = props;
 	const {
 		config: { isPaidContent, host, hasSurveyAd },
@@ -284,12 +282,6 @@ export const InteractiveLayout = (props: WebProps | AppsProps) => {
 							contributionsServiceUrl={contributionsServiceUrl}
 							showSlimNav={true}
 							showSubNav={false}
-							customSubnav={
-								props.NAV.customSubNav && {
-									data: props.NAV.customSubNav,
-									renderingPage: 'article',
-								}
-							}
 							hasPageSkin={false}
 							hasPageSkinContentSelfConstrain={false}
 							pageId={article.pageId}
@@ -304,7 +296,7 @@ export const InteractiveLayout = (props: WebProps | AppsProps) => {
 							<Section
 								fullWidth={true}
 								showTopBorder={false}
-								backgroundColour={sourcePalette.labs[100]}
+								backgroundColour={sourcePalette.labs[400]}
 								borderColour={sourcePalette.neutral[60]}
 								sectionId="labs-header"
 							>
@@ -319,7 +311,7 @@ export const InteractiveLayout = (props: WebProps | AppsProps) => {
 				</>
 			)}
 			<main data-layout="InteractiveLayout">
-				<DirectoryPageNavIsland
+				<DirectoryPageNav
 					pageId={article.pageId}
 					pageTags={article.tags}
 				/>
@@ -344,6 +336,7 @@ export const InteractiveLayout = (props: WebProps | AppsProps) => {
 										pageId={article.pageId}
 										webTitle={article.webTitle}
 										ajaxUrl={article.config.ajaxUrl}
+										abTests={article.config.abTests}
 										switches={article.config.switches}
 										isAdFreeUser={article.isAdFreeUser}
 										isSensitive={article.config.isSensitive}
@@ -493,15 +486,6 @@ export const InteractiveLayout = (props: WebProps | AppsProps) => {
 									)}
 								</div>
 							</GridItem>
-							{article.affiliateLinksDisclaimerRequired && (
-								<GridItem area="meta" element="aside">
-									<AffiliateDisclaimer
-										cssOverrides={css`
-											margin-top: ${space[4]}px;
-										`}
-									/>
-								</GridItem>
-							)}
 							<GridItem area="body" element="article">
 								<ArticleContainer format={format}>
 									<ArticleBody
@@ -511,6 +495,7 @@ export const InteractiveLayout = (props: WebProps | AppsProps) => {
 										pageId={article.pageId}
 										webTitle={article.webTitle}
 										ajaxUrl={article.config.ajaxUrl}
+										abTests={article.config.abTests}
 										switches={article.config.switches}
 										isSensitive={article.config.isSensitive}
 										isAdFreeUser={article.isAdFreeUser}
@@ -568,6 +553,9 @@ export const InteractiveLayout = (props: WebProps | AppsProps) => {
 									contributionsServiceUrl
 								}
 								idApiUrl={article.config.idApiUrl}
+								isMinuteArticle={
+									article.pageType.isMinuteArticle
+								}
 								isPaidContent={article.pageType.isPaidContent}
 								pageId={article.pageId}
 								sectionId={article.config.section}
@@ -618,6 +606,7 @@ export const InteractiveLayout = (props: WebProps | AppsProps) => {
 							article.showBottomSocialButtons &&
 							renderingTarget === 'Web'
 						}
+						isDeprecatedInteractiveLayout={true}
 					/>
 				</Section>
 
@@ -822,9 +811,11 @@ export const InteractiveLayout = (props: WebProps | AppsProps) => {
 							/>
 						</Island>
 					</BannerWrapper>
-					{renderAds && (
-						<MobileStickyContainer data-print-layout="hide" />
-					)}
+					<MobileStickyContainer
+						data-print-layout="hide"
+						contentType={article.contentType}
+						pageId={article.pageId}
+					/>
 				</>
 			)}
 			{isApps && (
