+++ dotcom-rendering/src/lib/renderElement.tsx
@@ -109,7 +109,6 @@ type Props = {
 	contentLayout?: string;
 	articleArrangement?: LayoutType;
 	idApiUrl?: string;
-	isShinyNewInteractiveLayout?: boolean;
 };
 
 // updateRole modifies the role of an element in a way appropriate for most
@@ -182,7 +181,6 @@ export const renderElement = ({
 	contentLayout,
 	articleArrangement,
 	idApiUrl,
-	isShinyNewInteractiveLayout = false,
 }: Props) => {
 	const isBlog =
 		format.design === ArticleDesign.LiveBlog ||
@@ -1056,7 +1054,6 @@ export const RenderArticleElement = ({
 	contentLayout,
 	articleArrangement,
 	idApiUrl,
-	isShinyNewInteractiveLayout = false,
 }: Props) => {
 	const withUpdatedRole = updateRole(element, format);
